extracted adding information from PortalInformation to request into own RequestProcessor
